+++ libs/velo-external-db-core/src/health/app_info.spec.ts
@@ -1,6 +1,6 @@
-const { Uninitialized, gen: genCommon } = require('@wix-velo/test-commons')
-const { appInfoFor, maskSensitiveData } = require ('./app_info')
-const driver = require('../../test/drivers/app_info_test_support') //TODO: change driver location
+import { Uninitialized, gen as genCommon } from '@wix-velo/test-commons'
+import { appInfoFor, maskSensitiveData } from './app_info'
+import * as driver from '../../test/drivers/app_info_test_support' //TODO: change driver location
 
 describe('App info Function', () => {
     test('get app info will retrieve valid config and will create app info object', async() => {
